# specs/002-account-auth-lifecycle/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Authentication and Account Lifecycle API
  version: 0.1.0
servers:
  - url: /v1
paths:
  /auth/sign-up:
    post:
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: Account created in unverified state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeAuthSession'
  /auth/sign-in:
    post:
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: Authentication success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeAuthSession'
        '401':
          description: Invalid credentials or blocked account state
  /auth/sign-out:
    post:
      summary: Sign out current session
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
  /auth/refresh:
    post:
      summary: Rotate refresh credential and return new access/refresh pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refresh success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeAuthSession'
        '401':
          description: Invalid, expired, revoked, or reused refresh credential
  /auth/verify-email:
    post:
      summary: Verify user email ownership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Verification success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
  /auth/resend-verification:
    post:
      summary: Resend verification challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
  /auth/password-reset/request:
    post:
      summary: Request password reset challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: Non-enumerating accepted response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
  /auth/password-reset/confirm:
    post:
      summary: Confirm password reset with challenge token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '200':
          description: Password reset success and prior sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
        '400':
          description: Invalid or expired challenge
  /auth/sessions:
    get:
      summary: List active sessions for current user
      responses:
        '200':
          description: Active session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeSessionList'
    delete:
      summary: Revoke all non-current sessions
      responses:
        '200':
          description: All non-current sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
  /auth/sessions/{sessionId}:
    delete:
      summary: Revoke one specific session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'

components:
  schemas:
    EnvelopeOk:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            ok:
              type: boolean
              enum: [true]
      required: [success, data]

    EnvelopeAuthSession:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/AuthSession'
      required: [success, data]

    EnvelopeSessionList:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: '#/components/schemas/SessionInfo'
      required: [success, data]

    SignUpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
          minLength: 2
          maxLength: 60
      required: [email, password, displayName]

    SignInRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]

    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required: [refreshToken]

    VerifyEmailRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required: [token]

    EmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]

    PasswordResetConfirmRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
      required: [token, newPassword]

    AuthSession:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/UserProfile'
      required: [accessToken, refreshToken, user]

    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        status:
          type: string
          enum: [active, suspended, deleted]
      required: [id, email, displayName, status]

    SessionInfo:
      type: object
      properties:
        id:
          type: string
        deviceLabel:
          type: string
        lastActivityAt:
          type: string
          format: date-time
        isCurrent:
          type: boolean
      required: [id, deviceLabel, lastActivityAt, isCurrent]